double on price MatUpdated + LogicFacade

+++ src/main/java/FunctionLayer/LogicFacade.java
@@ -30,7 +30,7 @@ public class LogicFacade
         return OrderMapper.createOrder(user, width, length, height, shed, status);
     }
 
-    public static void updateMaterial( String Name, int Length, int Stock, int Price, String Description, int MaterialsID) throws LoginSampleException {
+    public static void updateMaterial( String Name, int Length, int Stock, double Price, String Description, int MaterialsID) throws LoginSampleException {
     MaterialMapper.updateMaterial( Name, Length, Stock, Price, Description, MaterialsID);
     }
 }
